+++ .secrets.baseline
@@ -91,6 +91,10 @@
       "path": "detect_secrets.filters.allowlist.is_line_allowlisted"
     },
     {
+      "path": "detect_secrets.filters.common.is_baseline_file",
+      "filename": ".secrets.baseline"
+    },
+    {
       "path": "detect_secrets.filters.common.is_ignored_due_to_verification_policies",
       "min_level": 2
     },
@@ -176,15 +180,15 @@
         "line_number": 25
       }
     ],
-    "src/test/kotlin/com/noumenadigital/npl/cli/DeployCommandTestContainersIT.kt": [
+    "src/test/kotlin/com/noumenadigital/npl/cli/DeployCommandIT.kt": [
       {
         "type": "Secret Keyword",
-        "filename": "src/test/kotlin/com/noumenadigital/npl/cli/DeployCommandTestContainersIT.kt",
+        "filename": "src/test/kotlin/com/noumenadigital/npl/cli/DeployCommandIT.kt",
         "hashed_secret": "e38ad214943daad1d64c102faec29de4afe9da3d",
         "is_verified": false,
-        "line_number": 176
+        "line_number": 99
       }
     ]
   },
-  "generated_at": "2025-04-28T09:34:14Z"
+  "generated_at": "2025-04-28T14:24:09Z"
 }
