+++ src/test/kotlin/com/noumenadigital/npl/cli/TestUtils.kt
@@ -44,10 +44,6 @@ object TestUtils {
             .replace(Regex("in \\d+ ms"), "in XXX ms")
             // Remove any ANSI color codes
             .replace(Regex("\\e\\[[0-9;]*m"), "")
-            // Replace the random port number with a placeholder
-            .replace(Regex("localhost:(\\d{4,5})"), "localhost:XXXXX")
-            // Normalize the zip file name
-            .replace(Regex("zip:file:.+\\."), "zip:file:SOME_PATH.")
             .trimIndent()
 
     fun runCommand(
