--- src/test/resources/init-db-users.sql
@@ -1,3 +1,0 @@
-CREATE ROLE "platform_owner" LOGIN PASSWORD 'platform' NOINHERIT;
-GRANT CREATE ON DATABASE "platform" TO "platform_owner";
-CREATE EXTENSION pg_stat_statements;
