+++ src/test/resources/kotest.properties
@@ -1,0 +1,2 @@
+kotest.framework.classpath.scanning.config.disable=true
+kotest.framework.classpath.scanning.autoscan.disable=true
